Uodate the getCustomNormalizedFilePath option from a function to a resolved path
